float fix

+++ Learner.py
@@ -45,8 +45,8 @@ class DataLoaders:
 
             def get_window_score(idx): return float(subject.paired_windows[idx].score > 0)
 
-            X = torch.tensor([get_windows_avg_diff(idx) for idx in indices])
-            y = torch.tensor([get_window_score(idx) for idx in indices])
+            X = torch.tensor([get_windows_avg_diff(idx) for idx in indices]).float()
+            y = torch.tensor([get_window_score(idx) for idx in indices]).float()
 
             return DataLoader(TensorDataset(X, y), shuffle=False)
